feat: Enhance Flutter app UI and interactions
I've implemented several UI/UX improvements for your Flutter restaurant swiping application:

1.  **Login Screen:**
    *   The page transition to the main app screen now uses a fade effect instead of a slide.
    *   I've added a semi-transparent Tokyo cityscape image as the background.
    *   I've also improved the readability of the email/password fields against this new background.

2.  **Restaurant Swipe Screen:**
    *   The 'LIKE' button color is now yellowish-green.
    *   The 'NOPE' button color is now fluorescent red.
    *   I've added a bounce animation to the 'LIKE' and 'NOPE' buttons. This animation is triggered both by direct button presses and by swiping cards in the corresponding direction.

I validated these changes by running `flutter analyze` after setting up a basic project structure and `pubspec.yaml`.
